ath79: split DEVICE_TITLE, add DEVICE_RAM, add DEVICE_FLASH
This allows the definition of DEVICE_VENDOR, DEVICE_MODEL,
DEVICE_REVISION, DEVICE_RAM and DEVICE_FLASH. The first three can
be merged together to look the same as the current DEVICE_TITLE.
Also, if DEVICE_TITLE is set, it's directly used as a *fallback*
for devices which haven't been upgraded yet.

The motivation is to create JSON files based on provided metadata,
needing clear differentiation between vendor, model and revision.

Signed-off-by: Sudhanshu Gautam <me@sudhanshug.com>